fixed types

+++ rentacar/src/main/java/com/omertdemirel/rentacar/api/controller/CarController.java
@@ -1,4 +1,5 @@
 package com.omertdemirel.rentacar.api.controller;
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -59,7 +60,7 @@ public class CarController {
 	}
 
 	@GetMapping("/getByDailyPriceLessThanEqual")
-	public DataResult<List<ListCarDto>> getByDailyPriceLessThanEqual(double maxDailyPrice) {
+	public DataResult<List<ListCarDto>> getByDailyPriceLessThanEqual(BigDecimal maxDailyPrice) {
 		return carService.getByDailyPriceLessThan(maxDailyPrice);
 	}
